FIX (partial): sphinx_gallery_conf not pickleable
